Implement custom timer driver
Implements the necessary functions and traits for basic embassy time
driver. This custom driver uses the High-Resolution RTC (1kHz, also
called Wake) on the RT6xx chip. The RTC general purpose registers are
used to 1) expand the built-in 16 bit counter register to 32 bits and
pair it with a corresponding 32 bit match value for comparison and 2) to
function as an enable flag which will determine when to trigger the
embassy-time alarm.