Update iopctl interface to support GPIO
The GPIO HAL will set default slew rates and disable analog multiplexing
for all of its pins. To make this easier, combine "fail-safe" pin
interface into normal pin interface. If these settings are configured
for a high-speed pin, it will essentially be a no-op so there is
no danger in doing this.

Signed-off-by: Kurtis Dinelle <t-kudinelle@microsoft.com>