Create clock config structs and traits
This commit creates the clock config structures as well as a trait to
configure the relevant clock registers properly. There is a second trait for
clocks that may be configured with different sources which builds on top
of the ConfigurableClock trait. Finally, APIs are created and exposed
for the configuring clock out.