Remove DMA Instance from DMA and DMA-related structs' typestate
This commit removes the D: Instance type from the typestate of DMA,
Channel, and Transfer structs. This is a breaking change so the I2C
module constructors and the dma-mem example have been updated. The
DMA instance along with the Flexcomm instance will be removed from the
I2C structs in a later PR.