uart: pins are peripherals too
We want the Uart structure to maintain ownership of its pins such that
nobody else can instantiate any other object using the same pins. At
the same time, we can create RX-only or TX-only versions of the Uart,
which means both TX and RX are optional pins to the driver.