Remove typestate from I2C
This commit refactors the I2CMaster, I2CSlave, and I2CBus structs and their
implementations to remove both Flexcomm and DMA Instance from the
typestate. In place, the necessary Instances are passed as part of the
typestate of the constructors. All instance specific information is
stored in a <function>Info struct. i.e. DmaInfo. The I2C examples were
also updated to reflect these changes.