i2c: Remove Flexcomm typestate
This commit refactors the I2CMaster and I2CSlave structs and their
respective implementations to remove the Flexcomm Instance from the
typestate.

In its place, the necessary Instance is passed as part of the
typestate of the constructors.

All instance specific information is stored in a Info struct. The I2C
examples were also updated to reflect these changes.