uart: temporarily remove set_baudrate() API
We need to re-think how this API is supposed to be designed. First and
foremost, we shouldn't allow baudrate changes while UART is enabled
and we *always* enable UART in the constructors. Secondly, we want to
erase the Instance typestate and this method is creating challenges.

Considering the current implementation is unsound, let's remove it for
now.