dma: channel: skip abortctrl() field
Some changes being planned on the PAC will break this line. In
summary, we want to have per-channel bit fields in order to remove the
need for the unsafe usage. As a preparation for the changes in the
PAC, let's access the full register directly in order to avoid a build
break once the PAC is modified.