gpio: call gpio::init from HAL init()
Instead of exposing gpio::init(), call it from the HAL's init
function. With this change we can also assume that all GPIO interrupts
are masked at that point (reset state of the GPIO blocks) and remove
the use of unsafe as there's no way for users to call gpio::init()
from an invalid state.