i2c master write: fix dma + stop race condition
This is a race condition between DMA complete + sending the stop. DMA
completed ISR comes before last byte is acked, then we hit the stop bit
before it is ready.

It manifested in 2 similar ways that the SCL is held low for some time:

- stop() function exits before the last byte is even acked. We started
a new transaction with a restart and no stop to complete the previous
transaction. Sometimes the slave can tolerate this, but sometimes the
slave gets confused and gets into an error state.

- we hang in stop(), and the master state never returned to idle.

This add additional async wait after disabling DMA. Disabling DMA
re-enables the I2C master pending interrupt. If pending is set, that
means the last byte has been acked. If not, we wait for the pending
interrupt to let us know, it has been acked.

Also add check to beginning of stop() to make sure we are in the pending
state which means it is safe to hit the stop bit.